Add DataWriteSource interface

+++ include/Core/Util/Types.h
@@ -1,12 +1,14 @@
 #pragma once
 
+#include <cstdint>
+
 namespace ml {
 
 using byte = unsigned char;
 using sbyte = char;
 using uchar = unsigned char;
-using ushort = unsigned short;
-using uint = unsigned int;
-using ulong = unsigned long long;
+using ushort = uint16_t;
+using uint = uint32_t;
+using ulong = uint64_t;
 
 }
